refactor: common에 있는거 분리
