+++ app/_feature/components/Title.tsx
@@ -2,7 +2,7 @@ import { ArrowUturnLeftIcon } from "@heroicons/react/20/solid";
 import Link from "next/link";
 import { PropsWithChildren } from "react";
 
-import { classNames } from "@/common/utils/classNames";
+import { classNames } from "@/utils/classNames";
 
 interface TitleProps {
   showHomeLink?: boolean;
